fix: app init missing logger, ErrorUtils.setGlobalHandler and intl-pluralrules polyfill (#195)
### Description

The framework doesn't load `packages/@divvi/mobile/index.js`.
So it was deleted and the code was moved to
`packages/@divvi/mobile/src/app/App.tsx` instead.

### Test plan

- Tests pass

### Related issues

- N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)
